+++ src/oolib.nim
@@ -1,8 +1,11 @@
 import macros, sequtils
-import oolibpkg / [util]
-export optBase
+import oolibpkg / [sub, util]
+export optBase, pClass
 
-macro class*(head, body: untyped): untyped =
+macro class*(
+    head: untyped{nkIdent | nkCommand | nkInfix | nkCall | nkPragmaExpr},
+    body: untyped{nkStmtList}
+): untyped =
   let
     status = parseHead(head)
   var
@@ -13,29 +16,27 @@ macro class*(head, body: untyped): untyped =
     case node.kind
     of nnkVarSection:
       if status.kind == Alias:
-        error "Type Alias cannot have member variables", node
+        error "An alias class cannot have variables", node
       for n in node:
-        if n[^2].isEmpty:
-          # infer type from default
-          n[^2] = newCall(ident"typeof", n[^1])
+        n[^2] = n[^2] or newCall(ident"typeof", n[^1])
         argsList.add n
-    of nnkProcDef, nnkMethodDef, nnkFuncDef,
-      nnkIteratorDef, nnkConverterDef, nnkTemplateDef:
-      if node.kind == nnkProcDef:
-        cStatus.updateStatus(node)
-        if node.isConstructor: continue
-      elif node.kind == nnkMethodDef and status.kind == Inheritance:
+    of nnkProcDef:
+      cStatus.updateStatus(node)
+      if node.isConstructor: continue
+      result.add node.insertSelf(status.name)
+    of nnkMethodDef:
+      if status.kind == Inheritance:
         node.body = replaceSuper(node.body)
         result.add node.insertSelf(status.name).insertSuperStmt(status.base)
         continue
       result.add node.insertSelf(status.name)
+    of nnkFuncDef, nnkIteratorDef, nnkConverterDef, nnkTemplateDef:
+      result.add node.insertSelf(status.name)
     of nnkConstSection:
       for n in node:
-        if n[^2].isEmpty:
-          # infer type from default
-          n[^2] = newCall(ident"typeof", n[^1])
+        n[^2] = n[^2] or newCall(ident"typeof", n[^1])
         if n.last.isEmpty:
-          error "Consts must have a value", body
+          error "A constant must have a value", body
         constsList.add n
     of nnkDiscardStmt:
       return
@@ -54,15 +55,11 @@ macro class*(head, body: untyped): untyped =
   result[0][0][2][0][2] = argsList.map(delDefaultValue).toRecList()
 
 
-template pClass* {.pragma.}
-  ## Be used as pragma.
+proc isClass*(T: typedesc): bool =
+  ## Returns whether `T` is class or not.
+  T.hasCustomPragma(pClass)
 
 
 proc isClass*[T](instance: T): bool =
-  ## Whether `instance` is class or not.
-  T.hasCustomPragma(pClass)
-
-
-proc isClass*(T: typedesc): bool =
-  ## Whether `T` is class or not.
-  T.hasCustomPragma(pClass)
+  ## Is an alias for `isClass(T)`
+  T.isClass()